requirements.txt
flask==2.3.3
python-dotenv==1.0.0
flask-marshmallow==0.15.0
marshmallow==3.20.1
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.20
marshmallow-sqlalchemy==0.29.0
sqlalchemy-filters==0.13.0
Flask-Migrate==4.0.4
psycopg2==2.9.7
psycopg[binary]
